salesflow_auth/requirements.txt
# SalesFlow AI - Authentication Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Database
supabase>=2.0.0
httpx>=0.25.0

# Validation
pydantic>=2.5.0
pydantic[email]>=2.5.0
email-validator>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
python-dotenv>=1.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0

# Rate Limiting (optional - for production Redis-based)
# redis>=5.0.0
# aioredis>=2.0.0
